Extract locale dance to a concern
