+++ app/controllers/application_controller.rb
@@ -1,11 +1,14 @@
+require 'concerns/multi_lingualizable'
+
 class ApplicationController < ActionController::Base
+  include MultiLingualizable
+
   # TODO: comment captcha for ad creation/edition
   
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
-  before_action :set_locale
   before_action :configure_permitted_parameters, if: :devise_controller?
 
   rescue_from CanCan::AccessDenied do |exception|
@@ -27,39 +30,6 @@ class ApplicationController < ActionController::Base
     location_ask_path
   end
 
-  def set_locale
-    session[:locale] = params[:lang] if valid_locale?(params[:lang])
-
-    I18n.locale =
-      params_locale || session_locale || browser_locale || default_locale
-  end
-
-  def valid_locale?(code)
-    return false unless code.present?
-
-    I18n.available_locales.include?(code.to_sym)
-  end
-
-  def params_locale
-    params[:locale].presence
-  end
-
-  def session_locale
-    session[:locale].presence
-  end
-
-  def browser_locale
-    http_accept_language.compatible_language_from(I18n.available_locales)
-  end
-
-  def default_locale
-    I18n.default_locale
-  end
-
-  def self.default_url_options(options={})
-    { locale: I18n.locale }
-  end
-  
   def authenticate_active_admin_user!
     authenticate_user!
     unless current_user.admin?
